+++ repos/keg-components/src/theme/components/header/index.js
@@ -1,4 +1,4 @@
-import { baseHeader } from './baseHeader'
+import { itemHeader } from './itemHeader'
 export const header = {
-  baseHeader,
+  itemHeader,
 }
