+++ repos/keg-components/src/theme/components/header/itemHeader.js
@@ -37,7 +37,7 @@ const defaultSideSectionStyle = {
   },
 }
 
-export const baseHeader = {
+export const itemHeader = {
   main: {
     $all: {
       shadow: {
